Move prop components into separate components
