+++ src/components/HeroSection.js
@@ -1,8 +1,296 @@
-import React from "react";
+import React, {useState} from "react";
 import {Container} from "@material-ui/core"
 import Grid from "@material-ui/core/Grid"
+import {ERC721ABI} from "../ERC721ABI"
+import {
+  arweaveEndpoint,
+  getArweaveIDByCID,
+  getURLFromURI,
+  ipfsGetEndpoint,
+  knownGood,
+  knownPoor,
+  walkIPFSLinks
+} from "../utils"
+import Web3 from "web3"
+import checkMyNFT from "../images/logo.png";
+import eth from "../images/eth.png";
+import TextField from "@material-ui/core/TextField"
+import Button from "@material-ui/core/Button"
+import Alert from "@material-ui/lab/Alert"
+import Paper from "@material-ui/core/Paper"
 
-export default function HeroSection() {
+export default function HeroSection(props) {
+  let {
+    classes,
+    imageInfo,
+    setImageInfo,
+    nftAddress,
+    setNFTAddress,
+    tokenID,
+    setTokenID,
+    setNFTInfo,
+    errors,
+    setErrors,
+    fetchError,
+    setFetchError
+  } = props.props;
+
+  let wssOptions = {
+    timeout: 30000, // ms
+
+    clientConfig: {
+      // Useful if requests are large
+      maxReceivedFrameSize: 100000000, // bytes - default: 1MiB
+      maxReceivedMessageSize: 100000000, // bytes - default: 8MiB
+
+      // Useful to keep a connection alive
+      keepalive: true,
+      keepaliveInterval: 60000, // ms
+    },
+
+    // Enable auto reconnection
+    reconnect: {
+      auto: true,
+      delay: 5000, // ms
+      maxAttempts: 5,
+      onTimeout: false,
+    },
+  };
+
+  const [touched, setTouched] = useState({
+    nftAddress: false,
+    tokenID: false,
+  });
+
+  const [isLoading, setIsLoading] = useState(false);
+
+
+  const web3 = new Web3(
+      new Web3.providers.WebsocketProvider(
+          "wss://mainnet.infura.io/ws/v3/a30464df239144d0a8eae3f8a426d03e",
+          wssOptions
+      )
+  );
+
+  const validateTokenID = (tokenID) => {
+    if (+tokenID < 0) {
+      setErrors({ ...errors, tokenID: "TokenID cannot be negative" });
+      return;
+    }
+    setErrors({ ...errors, tokenID: "" });
+  };
+
+  const validateAddress = (address) => {
+    if (address.length === 40 || address.length === 42) {
+      setErrors({ ...errors, nftAddress: "" });
+      return;
+    }
+    setErrors({ ...errors, nftAddress: "Invalid address" });
+  };
+
+  const tryToGetTokenURI = async (contract, id) => {
+    let tokenURI = "";
+    let error = "";
+    try {
+      // this is proper ERC721 compatible
+      tokenURI = await contract.methods.tokenURI(id).call();
+    } catch (e) {
+      error = e.message;
+      try {
+        // this is NOT proper ERC721 but Rarible has this
+        tokenURI = await contract.methods.uri(id).call();
+        error = "";
+      } catch (e) {
+        error = e.message;
+      }
+    }
+    return [tokenURI, error];
+  };
+
+  const isButtonEnabled = () => {
+    return (
+        !touched.nftAddress ||
+        !touched.tokenID ||
+        errors.nftAddress !== "" ||
+        errors.tokenID !== ""
+    );
+  };
+
+  const handleClick = async () => {
+    setIsLoading(true);
+    try {
+      const contract = new web3.eth.Contract(ERC721ABI, nftAddress);
+
+      let owner = "";
+      try {
+        // this is proper ERC721 compatible
+        owner = await contract.methods.ownerOf(tokenID).call();
+      } catch (e) {
+        console.error(e);
+      }
+
+      const symbol = await contract.methods.symbol().call();
+      const name = await contract.methods.name().call();
+
+      if (knownPoor.includes(nftAddress.toLowerCase())) {
+        setNFTInfo({
+          level: "poor",
+          owner,
+          tokenURI: "N/A",
+          symbol,
+          name,
+          address: nftAddress,
+          tokenID: tokenID,
+          protocol: "N/A",
+          uriURL: "N/A",
+        });
+        // setImageInfo(defaultImgState); // todo check
+        setImageInfo(imageInfo);
+        return;
+      }
+      let [tokenURI, err] = await tryToGetTokenURI(contract, tokenID);
+      if (err !== "") {
+        console.error(err);
+        setFetchError("Could not fetch token URI for NFT " + tokenURI);
+        setIsLoading(false);
+        return;
+      }
+
+      let [uriURL, uriProtocol] = await getURLFromURI(tokenURI);
+
+      let uriResponse;
+      try {
+        uriResponse = await fetch(uriURL, { method: "GET" });
+      } catch (e) {
+        console.error(e);
+        setFetchError("Could not fetch NFT URI " + tokenURI);
+        setIsLoading(false);
+        return;
+      }
+
+      let uriInfo = await uriResponse.json();
+      let imgURI = uriInfo.image;
+
+      let [imageURIURL, protocol] = await getURLFromURI(imgURI);
+
+      let isImageOnArweave = false;
+      let isMetadataOnArweave = false;
+
+      // if the protocol is on IPFS, we check if the tokenURI AND the imageURI are stored on Arweave via ipfs2arweave.com
+      if (protocol === "ipfs") {
+        let metadataCID = uriURL.replace(ipfsGetEndpoint, "");
+        let rootMetadataCID = metadataCID.split("/");
+        if (rootMetadataCID.length !== 0) {
+          rootMetadataCID = await walkIPFSLinks(rootMetadataCID[0]);
+        }
+
+        let imageCID = imageURIURL.replace(ipfsGetEndpoint, "");
+        let rootImageCID = imageCID.split("/");
+        if (rootImageCID.length !== 0) {
+          rootImageCID = await walkIPFSLinks(rootImageCID[0]);
+        }
+
+        try {
+          // We check to see if the IPFS hash is stored on arweave
+          let arweaveImageID = await getArweaveIDByCID(rootImageCID);
+          if (arweaveImageID !== "") {
+            //  we change the imageURIURL to the arweaveImageID
+            imageURIURL = arweaveEndpoint + "/" + arweaveImageID;
+            isImageOnArweave = true;
+          }
+          let arweaveMetadataID = await getArweaveIDByCID(rootMetadataCID);
+          if (arweaveMetadataID !== "") {
+            tokenURI = arweaveEndpoint + "/" + arweaveMetadataID;
+            isMetadataOnArweave = true;
+          }
+        } catch (e) {
+          console.log(e);
+        }
+      }
+
+      setImageInfo({ ...imageInfo, loading: true });
+
+      fetch(imageURIURL, { method: "GET" })
+          .then(async (imageResponse) => {
+            let imageBlob = await imageResponse.blob();
+            let image = URL.createObjectURL(imageBlob);
+            setImageInfo({
+              imageURIURL: imageURIURL,
+              image: image,
+              loading: false,
+            });
+          })
+          .catch((e) => {
+            console.error(e);
+            setFetchError("Could not fetch NFT Image " + imgURI);
+            setIsLoading(false);
+          });
+
+      let severity = "undefined";
+      switch (uriProtocol) {
+        case "ipfs":
+          severity = "medium";
+          break;
+        case "arweave":
+          severity = "strong";
+          break;
+        case "centralized":
+          severity = "poor";
+          break;
+        default:
+          severity = "undefined";
+      }
+
+      if (knownGood.includes(nftAddress.toLowerCase())) {
+        setNFTInfo({
+          level: "strong",
+          owner,
+          tokenURI,
+          symbol,
+          name,
+          address: nftAddress,
+          tokenID: tokenID,
+          protocol: "On-chain",
+          uriURL,
+        });
+        setIsLoading(false);
+        return;
+      }
+      if (protocol === "ipfs" && isImageOnArweave && isMetadataOnArweave) {
+        setNFTInfo({
+          level: "strong",
+          owner,
+          tokenURI,
+          symbol,
+          name,
+          address: nftAddress,
+          tokenID: tokenID,
+          protocol: "Arweave & IPFS",
+          uriURL,
+        });
+        setIsLoading(false);
+        return;
+      }
+
+      setNFTInfo({
+        level: severity,
+        owner,
+        tokenURI,
+        symbol,
+        name,
+        address: nftAddress,
+        tokenID: tokenID,
+        protocol: uriProtocol,
+        uriURL,
+      });
+      setIsLoading(false);
+    } catch (e) {
+      console.error(e);
+      setFetchError(e.message);
+      setIsLoading(false);
+    }
+  };
+
   return (
       <Container>
         <Grid
